requirements.txt
Django==3.0.5
djangorestframework==3.10
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.2.1
drf-yasg==1.17.1
django-guardian==2.2.0
mkdocs==1.1
mkdocs-bootstrap4==0.1.4